datasources/github/requirements.txt
requests>=2.28.0
PyGithub>=1.58.0
markdown>=3.4.0
dify-plugins>=0.5.0b14